Reorganisation of init module
