+++ build.rs
@@ -5,7 +5,7 @@ fn main() {
     println!("cargo:rerun-if-changed=build.rs");
     println!("cargo:rerun-if-changed=src");
 
-    #[path = "src/init.rs"]
+    #[path = "src/init/mod.rs"]
     mod init;
 
     init::setup_shell(false);
